+++ R/exfm2.R
@@ -12,6 +12,7 @@
 #'   \item{STRATA_AREA}{area of each strata, in hectares}
 #'   \item{PLOT_AREA}{area of plots, in square meters}
 #'   \item{VWB}{volume with bark, in cubic meters}
+#'   \item{VWB_m3ha}{volume with bark, in cubic meters per hectare}
 #' }
 #' @name exfm2
 #' @docType data