+++ R/exfm5.R
@@ -11,6 +11,7 @@
 #'   \item{TOTAL_AREA}{total area, in hectares}
 #'   \item{PLOT_AREA}{area of plots, in square meters}
 #'   \item{VWB}{volume with bark, in cubic meters}
+#'   \item{VWB_m3ha}{volume with bark, in cubic meters per hectare}
 #' }
 #' @name exfm5
 #' @docType data